Refactor: GeocodingService 클래스명, 디렉토리 구조 리팩토링
